add apyfloat div-tests, re-organize unit tests
